requirements.txt
python-telegram-bot==20.7
openai==1.3.7
yt-dlp==2023.12.30
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
asyncio==3.4.3 